Refactor Duke.java, encapsulated code into Parser.java, Storage.java and TaskList.java
